Check for page updating
Used a mutation observer to monitor the source section for updates.
Sometimes this way is overkill and you could just do something like poll
the breadcrumbs for changes, but this is firing at just the right time
so sticking with it.
Also had to change wrapTextNodes but for the life of me cant remember
why ;)