Add highlights bars
Added the bar to the right (my fav bit;).
Also optimized the wraptextnodes a little and changed the css to use my
multiline string thingy now its got more css.
You had the css in another file and Im not against that at all, but for
dev reasons this was easier for me.  I did most my dev in the sources
section of chrome on a locally saved copy of the page, its quick and
easy for this sort of stuff.