Fix's for FireFox
Although I only code for Chrome (I only code for fun and its no fun
dealing with browser incompatibilities) I have a lot of love for FF and
the MDN so thought Id see if it worked.  Most of it did, it was just the
bar that had some problems, so fixed them.  You can now use this as a
GreaseMonkey script.